// Billing-Manager/types.ts
export interface Employee {
  id: string;
  fullName: string;
  role: string;
  hireDate: string; // ISO string YYYY-MM-DD
  salaryType: 'daily' | 'monthly';
  rate: number;
}

export interface EmployeeBenefit {
    id: string;
    employeeId: string;
    sss: boolean;
    philhealth: boolean;
    pagibig: boolean;
}

export interface TimeRecord {
    id: string;
    employeeId: string;
    date: string; // YYYY-MM-DD
    timeIn: string; // HH:MM
    timeOut: string; // HH:MM
}